+++ actionpack/CHANGELOG.md
@@ -1,3 +1,11 @@
+*   Removed deprecated AbstractController#skip_action_callback
+
+    *Iain Beeston*
+
+*   Removed deprecated *_filter callback methods from AbstractController
+
+    *Iain Beeston*
+
 *   Show an "unmatched constraints" error when params fail to match constraints
     on a matched route, rather than a "missing keys" error.